move files to correct path
