+++ src/main/kotlin/logic/usecases/project/ViewProjectHistoryUseCase.kt
@@ -1,7 +1,7 @@
-package logic.usecases
+package logic.usecases.project
 
 import logic.entities.LogItem
-import logic.exception.PlanMateException.ValidationException.InvalidProjectIDException
+import logic.exception.PlanMateException
 import logic.repository.LogRepository
 import java.util.UUID
 
@@ -9,12 +9,12 @@ class ViewProjectHistoryUseCase(
     private val logRepository: LogRepository
 ) {
     fun execute(projectId: String?): List<LogItem> {
-        if (projectId.isNullOrBlank()) throw InvalidProjectIDException
+        if (projectId.isNullOrBlank()) throw PlanMateException.ValidationException.InvalidProjectIDException
 
         val projectUUID = try {
             UUID.fromString(projectId)
         } catch (_: IllegalArgumentException) {
-            throw InvalidProjectIDException
+            throw PlanMateException.ValidationException.InvalidProjectIDException
         }
 
         return logRepository.viewLogsById(projectUUID)
